+++ app/src/lib/build-info.js
@@ -1,2 +1,2 @@
-export const APP_BUILD = 'v1.6.0'
-export const BUILD_DATE = '2026-06-07 16:00'
+export const APP_BUILD = 'v1.6.1'
+export const BUILD_DATE = '2026-06-07 17:30'